+++ MQGroup.PetShop.Domain/Services/PetService.cs
@@ -6,16 +6,19 @@ using System.Linq.Expressions;
 using MQGroup.PetShop.Core.IServices;
 using MQGroup.PetShop.Core.Models;
 using MQGroup.PetShop.Domain.IRepositories;
+using MQGroup.PetShop.Domain.Validators;
 
 namespace MQGroup.PetShop.Domain.Services
 {
     public class PetService : IPetService
     {
         private IPetRepository _repo;
+        private IValidator _validator;
 
-        public PetService(IPetRepository repo)
+        public PetService(IPetRepository repo, IValidator validator)
         {
             _repo = repo;
+            _validator = validator;
         }
         
         public List<Pet> GetAllPets()
@@ -25,21 +28,33 @@ namespace MQGroup.PetShop.Domain.Services
 
         public Pet CreatePet(Pet pet)
         {
+            if (!_validator.ValidatePet(pet))
+                throw new InvalidDataException(_validator.GetErrors());
+            
             return _repo.AddPet(pet);
         }
 
         public bool DeletePetById(int id)
         {
+            if (!_validator.PetExists(id))
+                throw new FileNotFoundException("Pet ID does not exist!");
             return _repo.DeletePetById(id);
         }
 
         public Pet GetPetById(int id)
         {
+            if (!_validator.PetExists(id))
+                throw new FileNotFoundException("Pet ID does not exist!");
             return _repo.GetPetById(id);
         }
 
         public Pet UpdatePet(int id, Pet pet)
         {
+            if (!_validator.PetExists(id))
+                throw new FileNotFoundException("Pet ID does not exist!");
+                
+            if (!_validator.ValidatePet(pet))
+                throw new InvalidDataException(_validator.GetErrors());
             return _repo.UpdatePet(id, pet);
         }
